separate out accordion row component
